fix: remove paperclip from migrations
some old migrations relied on recently removed paperclip gem.

+++ db/migrate/012_add_content_to_images_and_attachments.rb
@@ -1,16 +1,20 @@
 class AddContentToImagesAndAttachments < ActiveRecord::Migration[4.2]
 
-  TABLES = [:images, :attachments, :procurement_attachments]
+TABLES = [:images, :attachments, :procurement_attachments]
 
   def up
     TABLES.each do |table_name|
-      add_column table_name, :content, :text
+      if ActiveRecord::Base.connection.table_exists? table_name
+        add_column table_name, :content, :text
+      end
     end
   end
 
   def down
     TABLES.each do |table_name|
-      remove_column table_name, :content
+      if ActiveRecord::Base.connection.table_exists? table_name
+        remove_column table_name, :content
+      end
     end
   end
 end
